remove Domain and AbstractChunk types - move array domain stuff to darray.jl
