+++ src/Dagger.jl
@@ -7,17 +7,16 @@ include("lib/logging.jl")
 include("lib/dumbref.jl")
 
 # Data and slices of data
+include("thunk.jl")
 include("domain.jl")
-include("lib/blocked-domains.jl")
 include("chunks.jl")
 
 # Task scheduling
 include("processor.jl")
-include("thunk.jl")
 include("compute.jl")
 
 # Array computations
-include("array/lazy-array.jl")
+include("array/darray.jl")
 include("array/alloc.jl")
 include("array/map-reduce.jl")
